Books.html
Done with the filters and its backend part. We have to add in the database the themes for each book.

I have made a function checking if the image exists, but when checking that it gives an error in the console. I don't know how bad that is.

For the record I have used the ? operator. This works as an easy if. 
if (x) y else z is 
equal to
(x) ? y : z

LoadData is the javascript that will manage all the pages